Redundancy Fly Assembly as well as currently empty functions for wing manipulation and multiplying.
Commented out SetLevelFile in PlayingArea as it currently doesn't seem to work properly and only draws when a level is selected in. It seems it's meant to load all bugs for all levels beforehand, but currently will start level 0 with another level selected (if level 0 is selected, it plays level 0 twice. Any other level starts level 0 and the other level at the same time).